add issue presence as sign of project viability
It's a Good Thing when you run into a bug or missing feature and someone else already created an issue for it (or it is on the roadmap). Most issue trackers provide a means to indicate "me too".

I left these out of the "Bad for self-hosting" section because they felt too entitled:

* Issue tracker is missing a severe bug you're aware of.
* A severe issue has been logged but is ignored, disregarded, or triaged by maintainers.