WIP: remove dyndns and mail
mail isn't strictly necessary, so perhaps removing it will make mario simpler for beginners. The risk is this also makes mario less useful for more experienced self-hosters, but then again, it is a learning tool, not a true "homeprod" solution.

dyndns is complex because there are different ways to update records at the various dynamic DNS providers.