fix invalid HTML / just use links, not buttons
I can make a button work like a link, but it's a bit janky--the form action is nonobvious and doesn't preserve the query string (which is essential in this case).

See https://stackoverflow.com/questions/2906582/how-do-i-create-an-html-button-that-acts-like-a-link

If I care that much about looks I could do https://stackoverflow.com/questions/710089/how-do-i-make-an-html-link-look-like-a-button , but meh, links are just simpler and more straightforward UI/UX here.

And I don't think this is worth using JavaScript.