swap in valkey for redis
Valkey is a drop-in replacement with a proper FOSS license.

see https://help.nextcloud.com/t/using-valkey-as-a-drop-in-replacement-for-redis/202580

Easiest migration path is to first `dc SERVICE down` to destroy all `nextcloud` and `wallabag` containers before pulling this code, then start them back up.

Nextcloud also needs a change to its `config.php` file: change `redis` **where it appears as a hostname** to `valkey`.